Modify clear all logic such that only current user's comments are deleted

+++ portfolio/src/main/java/com/google/sps/servlets/DeleteServlet.java
@@ -28,6 +28,8 @@ import com.google.appengine.api.datastore.Key;
 import com.google.appengine.api.datastore.PreparedQuery;
 import com.google.appengine.api.datastore.Query;
 import com.google.appengine.api.datastore.Query.SortDirection;
+import com.google.appengine.api.users.UserService;
+import com.google.appengine.api.users.UserServiceFactory;
 import com.google.sps.data.RequestParameters;
 import com.google.sps.data.EntityProperties;
 import com.google.sps.servlets.AuthenticationServlet;
@@ -41,16 +43,24 @@ public class DeleteServlet extends HttpServlet {
     * is a specific comment id, that comment is deleted. */
   @Override
   public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
+    DatastoreService datastore = DatastoreServiceFactory.getDatastoreService();
+    UserService userService = UserServiceFactory.getUserService();
+    if (!userService.isUserLoggedIn()) {
+      return;
+    } 
+    String currentUserId = userService.getCurrentUser().getUserId();
+
     String whichCommentToDelete = 
       (String) request.getParameter(RequestParameters.WHICH_COMMENT_TO_DELETE);
-
-    DatastoreService datastore = DatastoreServiceFactory.getDatastoreService();
 
     if (whichCommentToDelete.equals("\"all\"")) {
       Query query = new Query("Comment");
       PreparedQuery results = datastore.prepare(query);
       results.asIterable().forEach(comment -> {
-        datastore.delete(comment.getKey());
+        if (currentUserId.equals((String) comment.getProperty(
+          EntityProperties.USER_ID_OF_AUTHOR))) {
+          datastore.delete(comment.getKey());
+        }
       });
     } else {
       long id = Long.parseLong(whichCommentToDelete);
